requirements.txt
lxml==3.5.0
requests==2.9.1
transmission-fluid==0.6
